requirements.txt
functions-framework==3.5.0
google-generativeai>=0.3.2
flask>=3.0.0
gunicorn==21.2.0
